requirements.txt
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0
--index-url https://download.pytorch.org/whl/cu121
transformers>=4.42.0
diffusers>=0.30.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
peft
scipy>=1.13.0
sentence-transformers
spacy
streamlit
fastapi
uvicorn[standard]
pymupdf
pillow
opencv-python
python-multipart
huggingface-hub
einops
timm
protobuf